requirements.txt
python-telegram-bot==12.2.0
requests==2.22.0
schedule==0.6.0
